+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJMVZPRGTETHE6T3PJXCJ38J"
+        "01KJPZM9R8ZQW1FPK87P1XBYNE"
     )
     r"""The unique identifier of the data source"""